Se corrigio un error de escritura en nombre del proyecto, se actualizo a Lootbyte, también se creo los menús correspondientes para el cliente y para el administrador y por ultimo se creo un header con una barra de busqueda, el icono de corazón y unos tab de filtrado de productos, como también se realizounos headers simples que tienen un icono de flecha para devolverse y queda pendiente el titulo de cada activity
