+++ app/build.gradle.kts
@@ -3,7 +3,7 @@ plugins {
 }
 
 android {
-    namespace = "com.example.lookbyte"
+    namespace = "com.example.lootbyte"
     compileSdk {
         version = release(36) {
             minorApiLevel = 1
@@ -11,7 +11,7 @@ android {
     }
 
     defaultConfig {
-        applicationId = "com.example.lookbyte"
+        applicationId = "com.example.lootbyte"
         minSdk = 25
         targetSdk = 36
         versionCode = 1
