+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.LookByte">
+        android:theme="@style/Theme.LootByte">
+        <activity
+            android:name=".UI.MainMenu.Admin.AdminActivity"
+            android:exported="false" />
         <activity
             android:name=".MainActivity"
             android:exported="true">
