+++ app/src/main/java/com/example/lootbyte/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.lookbyte
+package com.example.lootbyte
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -8,10 +8,10 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
-import com.example.lookbyte.UI.MainMenu.Cliente.CarritoFragment
-import com.example.lookbyte.UI.MainMenu.Cliente.HomeFragment
-import com.example.lookbyte.UI.MainMenu.Cliente.OfertasFragment
-import com.example.lookbyte.UI.MainMenu.Cliente.PerfilFragment
+import com.example.lootbyte.UI.MainMenu.Cliente.CarritoFragment
+import com.example.lootbyte.UI.MainMenu.Cliente.HomeFragment
+import com.example.lootbyte.UI.MainMenu.Cliente.OfertasFragment
+import com.example.lootbyte.UI.MainMenu.Cliente.PerfilFragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
@@ -25,11 +25,11 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            v.setPadding(systemBars.left, 0, systemBars.right, systemBars.bottom)
             insets
         }
 
-        val toobal = findViewById<Toolbar>(R.id.toolbar)
+//        val toobal = findViewById<Toolbar>(R.id.toolbar)
         drawerLayout= findViewById(R.id.main)
         val bottomNav = findViewById<BottomNavigationView>(R.id.bottom_nav)
