+++ app/src/main/java/com/example/lootbyte/UI/MainMenu/Cliente/CarritoFragment.kt
@@ -1,11 +1,11 @@
-package com.example.lookbyte.UI.MainMenu.Admin
+package com.example.lootbyte.UI.MainMenu.Cliente
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.lookbyte.R
+import com.example.lootbyte.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ReportesFragment.newInstance] factory method to
+ * Use the [CarritoFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ReportesFragment : Fragment() {
+class CarritoFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class ReportesFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_reportes, container, false)
+        return inflater.inflate(R.layout.fragment_carrito, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class ReportesFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ReportesFragment.
+         * @return A new instance of fragment CarritoFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ReportesFragment().apply {
+            CarritoFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
