+++ app/src/main/java/com/example/lootbyte/UI/MainMenu/Cliente/HomeFragment.kt
@@ -1,11 +1,12 @@
-package com.example.lookbyte.UI.MainMenu.Admin
+package com.example.lootbyte.UI.MainMenu.Cliente
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.lookbyte.R
+import com.example.lootbyte.R
+import com.google.android.material.tabs.TabLayout
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -14,10 +15,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [PerfilFragment.newInstance] factory method to
+ * Use the [HomeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class PerfilFragment : Fragment() {
+class HomeFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +36,18 @@ class PerfilFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_perfil2, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val tabLayout = view.findViewById<TabLayout>(R.id.tab_categorias)
+
+        tabLayout.addTab(tabLayout.newTab().setText("Todo"))
+        tabLayout.addTab(tabLayout.newTab().setText("Mouse"))
+        tabLayout.addTab(tabLayout.newTab().setText("Teclados"))
+        tabLayout.addTab(tabLayout.newTab().setText("Pantallas"))
+        tabLayout.addTab(tabLayout.newTab().setText("Audífonos"))
     }
 
     companion object {
@@ -45,12 +57,12 @@ class PerfilFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment PerfilFragment.
+         * @return A new instance of fragment HomeFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            PerfilFragment().apply {
+            HomeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
