+++ app/src/main/res/drawable/box_solid_full.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="640dp"
+    android:height="640dp"
+    android:viewportWidth="640"
+    android:viewportHeight="640">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M465.4,192L431.1,144L209,144L174.7,192L465.4,192zM96,212.5C96,199.2 100.2,186.2 107.9,175.3L156.9,106.8C168.9,90 188.3,80 208.9,80L431,80C451.7,80 471.1,90 483.1,106.8L532,175.3C539.8,186.2 543.9,199.2 543.9,212.5L544,480C544,515.3 515.3,544 480,544L160,544C124.7,544 96,515.3 96,480L96,212.5z"/>
+</vector>
