+++ app/src/main/res/drawable/busqueda.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid
+        android:color="@color/busqueda_background"/>
+    <corners
+        android:radius="30dp"/>
+
+</shape>
