+++ app/src/main/res/drawable/tab_indicator.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/black" />
+    <corners android:radius="4dp" />
+</shape>
