+++ app/src/main/res/layout/fragment_home.xml
@@ -5,10 +5,16 @@
     android:layout_height="match_parent"
     tools:context=".UI.MainMenu.Cliente.HomeFragment">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <include layout="@layout/header_busqueda" />
+    <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:layout_height="match_parent">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="20dp"/>
+    </androidx.core.widget.NestedScrollView>
+
 
 </FrameLayout>
