+++ app/src/main/res/menu/menu_bottom_nav_admin.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_producto"
+        android:title="@string/producto"
+        android:icon="@drawable/book_solid_full"/>
+
+    <item
+        android:id="@+id/nav_usuarios"
+        android:title="@string/usuarios"
+        android:icon="@drawable/users_solid_full"/>
+    <item
+        android:id="@+id/nav_pedidos"
+        android:title="@string/pedidos"
+        android:icon="@drawable/box_solid_full"/>
+    <item
+        android:id="@+id/nav_perfil_admin"
+        android:title="@string/perfil"
+        android:icon="@drawable/user_regular_full"/>
+</menu>
