+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,5 @@
     <color name="nav_morado_activo">#8F29085C</color>
     <color name="Texto_busqueda">#CFCBCB</color>
     <color name="busqueda_background">#4C1D95</color>
+    <color name="tabs_filtro">#DBDECD</color>
 </resources>
