+++ app/src/main/res/values/themes.xml
@@ -1,11 +1,25 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.LookByte" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.LootByte" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.LookByte" parent="Base.Theme.LookByte">
+    <style name="Theme.LootByte" parent="Base.Theme.LootByte">
         <item name="colorPrimary">@color/ic_launcher_background</item>
+    </style>
+
+<!--    header simple-->
+    <style name="HeaderShape">
+        <item name="cornerFamily">rounded</item>
+        <item name="cornerSizeTopLeft">0dp</item>
+        <item name="cornerSizeTopRight">0dp</item>
+        <item name="cornerSizeBottomLeft">24dp</item>
+        <item name="cornerSizeBottomRight">24dp</item>
+    </style>
+
+    <!--    header busqueda-->
+    <style name="HeaderBusqueda">
+
     </style>
 </resources>
